pyproject.toml
[tool.poetry]
name = "bushfire-beacon"
version = "0.1.0"
description = "simple bushfire beacon"
authors = ["Dax Reyes <dax.reyes@woohoow.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.65.2"}
python-jose = "^3.3.0"
SQLAlchemy = "^1.4.18"
alembic = "^1.6.5"
loguru = "^0.5.3"
pydash = "^5.0.0"
SQLAlchemy-Utils = "^0.37.7"
passlib = "^1.7.4"
emails = "^0.6"
phonenumbers = "^8.12.25"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
